Aquality2/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
config==0.5.0.post0
Django==3.1.3
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-gis==0.16
drf-extra-fields==3.0.4
idna==2.10
importlib-metadata==3.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.0
python-decouple==3.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
zipp==3.4.0
